% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia.R
\name{eia_query}
\alias{eia_query}
\title{EIA API Category Query}
\usage{
eia_query(api_key, category_id = NULL)
}
\arguments{
\item{api_key}{A character, the user API key for the eia website}

\item{category_id}{A character, the category ID as defined in the eia API}
}
\value{
A list, with the series metadata
}
\description{
A function to pull the available categories in the eia API
}
