% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia.R
\name{eia_series}
\alias{eia_series}
\title{EIA API Series Query}
\usage{
eia_series(api_key, series_id)
}
\arguments{
\item{api_key}{A character, the user API key for the eia website}

\item{series_id}{A character, the series ID as defined in the eia API,
to query for the available series IDs use the eia_query function}
}
\value{
A list, with the series metadata
}
\description{
A function to pull a series from the eia API based on series ID
}
