% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia.R
\name{eia_parse}
\alias{eia_parse}
\title{Parse a EIA Output}
\usage{
eia_parse(raw_series, type = "xts")
}
\arguments{
\item{raw_series}{A list, the EIA API output for series request using the eia_series function}

\item{type}{A character, define the class of the output, possible options c("xts", "zoo", "ts", "data.frame", "data.table", "tbl")}
}
\value{
A time series object according to the type argument setting
}
\description{
A parsing function for series from the EIA API
}
