% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia.R
\name{eia_query}
\alias{eia_query}
\title{EIA API Category Query}
\usage{
eia_query(api_key, category_id = NULL)
}
\arguments{
\item{api_key}{A character, the user API key for the eia website}

\item{category_id}{A character, the category ID as defined in the eia API}
}
\value{
A list, with the series metadata
}
\description{
A function to pull the available categories in the eia API
}
\examples{
\dontrun{
# Set you eia API key
api_key <- "Set you API key"

# Querying the eia API to get possible categories
eia_query(api_key = api_key) # getting the full list of categories
eia_query(api_key = api_key, category_id = "0") # querying the Electricity category
childseries <- eia_query(api_key = api_key, category_id = "1") # querying the Electricity Net Generation child series
}
}
