% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia.R
\name{eia_series}
\alias{eia_series}
\title{EIA API Series Query}
\usage{
eia_series(api_key, series_id)
}
\arguments{
\item{api_key}{A character, the user API key for the eia website}

\item{series_id}{A character, the series ID as defined in the eia API,
to query for the available series IDs use the eia_query function}
}
\value{
A list, with the series metadata
}
\description{
A function to pull a series from the eia API based on series ID
}
\examples{
\dontrun{
# Set you eia API key
api_key <- "Set you API key"

# Querying the eia API to get possible categories
eia_query(api_key = api_key) # getting the full list of categories
eia_query(api_key = api_key, category_id = "0") # querying the Electricity category
childseries <- eia_query(api_key = api_key, category_id = "1") # querying the Electricity Net Generation child series
series_id <- childseries$category$childseries$series_id[1] # pulling the first series ID from the list of Net Generation series

# Pulling the series from the eia API
raw_series <- eia_series(api_key = api_key, series_id = series_id)
}
}
